Add controller to pay fees and display fee reports

+++ clients/src/main/java/com/sas/clients/academic_terms/AcademicYearTermResponse.java
@@ -13,7 +13,7 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AcademicYearTermResponse {
-    private Integer yearId;
+    private Long yearId;
     private Integer termId;
     private String year;
     private String term;
